Updated Server code, Added Message class and Enum
Updated the server class so that it uses object streams and reads from them. Includes several if statements to handle each of the different message types but the functionality for the branches is missing. Included empty methods for each of the branches as well and comments describing what each one should do. Client code has the socket functionality and object streams but they are commented out as I was testing the GUI aspect of the project.