Full Command line functionality implemented.
The client and server applications now have full functionality for the altering of user data. This means that the client has several examples for sending messages to the server and the server's response to said messages such as validating user information (log in credentials) and updating their balance. This is all in command line however, will work on GUI implementation now. Additionally, I have tested the multithreaded server by connecting to it from two client applications on different devices and interacting with the server at the same time through updateBalance(). The server successfully receives the request from both client applications and updates the userData.txt file on the server's end.