View Profile GUI Window implemented
The view profile button for the main menu GUI window operates as it should, sending a message to the server to which the server should reply with a message containing the user's data. Originally, when creating a new user and then pressing view profile, there was a bug that occurred, freezing the entire application but that has been resolved. Will work on the presentation slides now